move list command to archiver.list
